add: table modal for each appointment
